Tweaked GA, Added report

+++ Assignment2/Genetic Algorithm/GA/GenerationLogger.cpp
@@ -6,18 +6,18 @@ void GenerationLogger::Log( const std::vector<Unit*>& Generation ) {
 	std::ofstream File;
 
 	File.open( "Logs/Median.txt", std::ios_base::app );
-	File << std::to_string( Generation[Generation.size()/2]->FitnessFunction() ) << std::endl;
+	File << std::to_string( Generation[Generation.size() / 2]->FitnessFunction() ) << std::endl;
 	File.close();
 
 	File.open( "Logs/Average.txt", std::ios_base::app );
-	int Average = 0;
+	float Average = 0;
 	{
 		for each (auto Unit in Generation) {
 			Average += Unit->FitnessFunction();
 		}
 		Average /= Generation.size();
 	}
-	File << std::to_string(Average) << std::endl;
+	File << std::to_string( Average ) << std::endl;
 	File.close();
 
 	File.open( "Logs/Best.txt", std::ios_base::app );
@@ -25,7 +25,7 @@ void GenerationLogger::Log( const std::vector<Unit*>& Generation ) {
 	File.close();
 
 	File.open( "Logs/Worst.txt", std::ios_base::app );
-	File << std::to_string( Generation[Generation.size() - 1 ]->FitnessFunction() ) << std::endl;
+	File << std::to_string( Generation[Generation.size() - 1]->FitnessFunction() ) << std::endl;
 	File.close();
 
 	File.open( "Logs/StatsBest.txt", std::ios_base::app );
